include actions in total reservation cost calculation, refine UI for training dates availability

+++ FitToFit/UI/fittofit_mobile/lib/models/rezervacijeRequest.dart
@@ -2,7 +2,7 @@ class RezervacijeRequest {
   String datum;
   int korisnikId;
   int clanarinaId;
-  int? iznos;
+  double? iznos;
   String datumIsteka;
   List<RezervacijeItem> items;
